+++ src/main/java/com/cetin/userpostapi/service/ProductService.java
@@ -4,6 +4,9 @@ import com.cetin.userpostapi.dto.ProductDto;
 import com.cetin.userpostapi.entity.Product;
 import com.cetin.userpostapi.repository.IProductRepository;
 import org.modelmapper.ModelMapper;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -16,34 +19,40 @@ public class ProductService {
     private final IProductRepository productRepository;
     private final ModelMapper modelMapper;
 
-
+    @Autowired
     public ProductService(IProductRepository productRepository, ModelMapper modelMapper) {
         this.productRepository = productRepository;
         this.modelMapper = modelMapper;
     }
 
-    public void saveProduct(ProductDto productDto) {
+    public ResponseEntity<Void> saveProduct(ProductDto productDto) {
         Product product = modelMapper.map(productDto, Product.class);
         productRepository.save(product);
+        return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
-    public List<ProductDto> getAllProducts(){
+    public ResponseEntity<List<ProductDto>> getAllProducts() {
         List<Product> products = productRepository.findAll();
-        return products.stream()
-                .map(product -> modelMapper.map(product,ProductDto.class))
+        List<ProductDto> productDtos = products.stream()
+                .map(product -> modelMapper.map(product, ProductDto.class))
                 .collect(Collectors.toList());
+        return new ResponseEntity<>(productDtos, HttpStatus.OK);
     }
 
-    public Optional<ProductDto> getProductById(Long id){
+    public ResponseEntity<ProductDto> getProductById(Long id) {
         Optional<Product> optionalProduct = productRepository.findById(id);
-        return optionalProduct
-                .map(product -> modelMapper.map(product,ProductDto.class));
+        return optionalProduct.map(product -> new ResponseEntity<>(modelMapper.map(product, ProductDto.class), HttpStatus.OK))
+                .orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
     }
 
-    public void deleteProduct(Long id){
-        Optional<Product> optionalProduct =productRepository.findById(id);
-        productRepository.delete(optionalProduct.get());
+    public ResponseEntity<Void> deleteProduct(Long id) {
+        Optional<Product> optionalProduct = productRepository.findById(id);
+
+        if (optionalProduct.isPresent()) {
+            productRepository.delete(optionalProduct.get());
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
-
-
 }
